Run ava test serially
DashboardRoute tests

GeneralRoute tests

SourceRoute tests
