+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "cross-env NODE_ENV=production node src/index.js",
     "dev": "cross-env NODE_ENV=development nodemon src/index.js",
-    "test": "c8 ava",
+    "test": "c8 ava --serial",
     "lint": "./node_modules/.bin/eslint src"
   },
   "author": "",
